lich su mua hang, them sp tu detail sang view detail

+++ src/main/java/vn/smartapple/appleshop/domain/CartDetail.java
@@ -24,7 +24,7 @@ public class CartDetail {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     long id;
 
-    int quantity;
+    long quantity;
 
     double price;
